dispatch drag enter and leave events
When a draggable is first dragged over a dropzone, an interactdragenter event is dispatched to the dropzone element.
If the draggable was previously being dragged over another dropzone, then dispatch an interactdragleave event to the previous dropzone.
Both these events have an event.detail.draggable property which is the element is being dragged.