Add preventDefault and drag axis options
preventDefault
--------------

Returns or sets whether to prevent the browser's default behaviour in
response to pointer events. Can be set to

- true to always prevent
- false to never prevent
- 'auto' to allow interact.js to try to guess what would be best
- null to set to the default ('auto')

drag axis
---------

interact.drag({ axis: 'x' || 'y' || 'xy' })

If set to 'x' or 'y', then the drag event sequence will only happen if
the direction of the first movement of the drag is along the given axis.