Save duplicate of the pointer down event
On iOS it seems that the touch objects in event.touches are the same
objects from touchstart to touchend of a point of contact.

Because an action doesn't start until the pointer moves, there is a
difference between the coordinates saved in 'startCoords' and those that
are reported by the touchstart events's touch coordinates. This
difference caused an incorrect startOffset to be calculated for
snapping/restricting when elementOrigin/elementRect are used.

Closes https://github.com/taye/interact.js/issues/73