clean up module implementation by removing conditional exports
Again a subjective change, but by having exports statements in conditional blocks it makes it unclear what values are being exported by the module, making it necessary to read all the code to get a grasp of what is happening.  Effectively, exporting inside conditionals is adding cyclomatic complexity to the module definition, its much clearer to the reader of the module if they can look in a single literal and see what they can expect to get from the module.

Cleaned up the Shadow DOM check to make it more self explanatory