Fix calculation of angle, Change which finger is the origin when calculating the angle.
Calculate the angle using atan2 instead of atan. This lets the angle cover the full circle and be used in
downstream calculations.

Calculate the angle from the first finger placed to the second, rather than the reverse as previously done.
This is important to the user when they use a finger from each hand for large rotations in rotate-scale
operations.