Fix aliased default options
Because extend does a shallow copy of object properties, some
options in interactable objects were aliased to objects in
the default objects. In some cases, setting options on one
interactable would bleed over to interactables made subsequently.

This commit adds a clone utility which is similar to extend,
but makes new objects instead of aliasing them. This allow the
newly added tests to pass.